requirements.txt
Django~=6.0.1
djangorestframework~=3.16.1
django-filter~=25.2
openpyxl~=3.1.5
pillow~=12.1.0